refactor: fix InputSizeType import
